requirements.txt
absl-py
kubernetes
